TRAD-82 Multiple Changes
Limited Bullet Count to 500 In the Inspector, any higher than that and we were hitting pretty big performance losses.
Temporary thing to make both states transition to each other for testing purposes
Update to ObjectPool, I made a mistake in all methods, fixed it in one and then didn't apply to the rest. So applying to all of them here